requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.4
pydantic-settings==2.6.1
boto3==1.35.68
httpx==0.27.2
alembic==1.14.0
traceloop-sdk==0.49.8
opentelemetry-sdk>=1.27.0,<2
opentelemetry-exporter-otlp>=1.27.0,<2
opentelemetry-instrumentation-fastapi>=0.48b0,<1
opentelemetry-instrumentation-sqlalchemy>=0.48b0,<1

